requirements.txt
cloudshell-automation-api>=9.0.0.0,<9.0.1.0
cloudshell-shell-core>=5.0.3,<6.0.0
cloudshell-shell-standards~=1.2
cloudshell-cp-core~=2.3
cloudshell-logging>=1.0.0,<1.1.0
cloudshell-shell-connectivity-flow~=3.0
lxml==4.6.1
requests==2.18.4
pyvmomi~=7.0.0
jsonpickle~=1.4
retrying==1.3.3
cryptography>=2.5,<3.3
pyOpenSSL<20.0.0
packaging>=20.9
netaddr==0.8.0
jsonschema==4.1.0
attrs~=21.0
